refactor: split window resolution tests into a sibling module
